change event creation process and popover functionality
Remove `Add Event` button and make modal appear after the selection.
Add a help button to show a new modal with information about what the user can do in this page.
Up to 2 events/tasks can be rendered in the same time slot in the week view, if there are more, they will be a small block that link to the day view of that specific day.